requirements.txt
streamlit==1.18.1
opencv-python==4.5.5.64
numpy==1.21.6
Pillow==9.0.1
